refactor: clean up seed endpoint logic
