+++ backend/app/schemas/seed.py
@@ -5,5 +5,4 @@ class SeedDatasetResponse(BaseModel):
     ok: bool
     record_count: int
     inserted_count: int = 0
-    sample_address: str | None = None
     message: str
